+++ py-rattler/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.1] - 2026-03-10
+
+### Added
+
+- Add methods to download a file by @pavelzw in [#2201](https://github.com/conda/rattler/pull/2201)
+
 ## [0.23.0] - 2026-03-06
 
 ### Added
@@ -18,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement comparison and hashing for `PackageRecord` types ([#2046](https://github.com/conda/rattler/pull/2046))
 - Support glob and regex patterns in repodata queries ([#2036](https://github.com/conda/rattler/pull/2036))
 - Add OAuth/OIDC authentication support in the authentication middleware ([#2049](https://github.com/conda/rattler/pull/2049))
+- Add extra to AboutJson ([#2198](https://github.com/conda/rattler/pull/2198))
 
 ### Changed
 
@@ -52,5 +59,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimized repodata loading: up to 65x faster for in-memory queries and 2.4x faster for warm disk cache ([#2058](https://github.com/conda/rattler/pull/2058))
 - Speed up matchspec parsing by ~2x ([#2066](https://github.com/conda/rattler/pull/2066))
 
-[Unreleased]: https://github.com/conda/rattler/compare/py-rattler-v0.23.0...HEAD
+[Unreleased]: https://github.com/conda/rattler/compare/py-rattler-v0.23.1...HEAD
 [0.23.0]: https://github.com/conda/rattler/compare/py-rattler-v0.22.0...py-rattler-v0.23.0
+[0.23.1]: https://github.com/conda/rattler/compare/py-rattler-v0.23.0...py-rattler-v0.23.1